renamed:    Input/Hg1201-tabis.txt -> Input/data/Hg1201-tabis.txt 	renamed:    Input/LBCO-hucker.txt -> Input/data/LBCO-hucker.txt 	renamed:    Input/YBCO-blanco.txt -> Input/data/YBCO-blanco.txt 	renamed:    Input/YBCO-hucker.txt -> Input/data/YBCO-hucker.txt 	new file:   Input/dataHg1201-tabis.txt 	new file:   Input/manifest 	modified:   Makefile 	new file:   SCStripe 	new file:   resultados.dat 	new file:   src/IO/IO.f03 	new file:   src/IO/create_input_manifest.f03 	renamed:    src/crystal/find_in_file.f03 -> src/IO/find_in_file.f03 	new file:   src/IO/teste.f03 	renamed:    src/m.f03 -> src/UI/m.f03 	modified:   src/analysis/analysis.f03 	modified:   src/analysis/export_analysis.f03 	modified:   src/analysis/standartPrint.f03 	modified:   src/crystal/crystalbase.f03 	modified:   src/crystal/setCrystal.f03 	modified:   src/main.f03 	renamed:    Input/styles/hg_tabis_energyGap.gp -> styles/hg_tabis_energyGap.gp 	renamed:    Input/styles/lbco_hucker_energyGap.gp -> styles/lbco_hucker_energyGap.gp 	renamed:    Input/styles/ybco_blanco_energyGap.gp -> styles/ybco_blanco_energyGap.gp 	renamed:    Input/styles/ybco_hucker_energyGap.gp -> styles/ybco_hucker_energyGap.gp

+++ src/crystal/crystalbase.f03
@@ -17,12 +17,12 @@ module crystalBase
         integer                        :: dataLEN                         !Número de pontos.
         real                           :: charge_mod                      !Modulação de de carga dentro do poço ()
         real                           :: unit_cell_100                   !Distancia [100] do cristal
-        character(:),allocatable          :: name                            !Nome do cristal
+        character(:),allocatable       :: name                            !Nome do cristal
     !-----------------------------------------------------------------------------------------------------------------------!
         real, allocatable                :: corr_L(:), ee_corr_L(:)         !Comprimento de correlação ; Erro sobre corr_L 
         real, allocatable                :: dope(:)                         !Dopagem do cristal (hole/unity cell)
         real*8, allocatable              :: E_0(:)                          !Parâmetro E_0
-        integer(kind=2), allocatable     :: Nmax(:)                         !Nível de preenchimento dos elétrons sobre o poço quadrado
+        integer, allocatable             :: Nmax(:)                         !Nível de preenchimento dos elétrons sobre o poço quadrado
     !--------------------------------------------------------------------------------------------------------------------------------
         
         type(Earray), dimension(:), allocatable               :: well_E       !Array com as energias do poço quadrado com Comprimento 'corr_L'
@@ -56,7 +56,7 @@ contains
 
 #include<../src/crystal/setParams.f03>
 
-#include<../src/crystal/find_in_file.f03>
+#include<../src/IO/find_in_file.f03>
 
 #include<../src/crystal/printCrystalData.f03>
